+++ docs/en/changes/changes.md
@@ -48,6 +48,7 @@
 * Replace Metrics v2 protocol with MQE in UI templates and E2E Test.
 * Fix incorrect apisix metrics otel rules.
 * Support `Scratch The OAP Config Dump`.
+* Support `increase/rate` function in the `MQE` query language.
 * Group service endpoints into `_abandoned` when endpoints have high
   cardinality.